@@ -8,7 +8,6 @@ class Study < ActiveRecord::Base
   has_role(:lab_manager)
 
   json do
-
     store_as_boolean(
       :contains_human_dna,
       :contaminated_human_dna
@@ -21,10 +20,10 @@ class Study < ActiveRecord::Base
     )
 
     translate(
-      :id                => :id_study_lims,
-      :uuid              => :uuid_study_lims,
-      :sac_sponsor       => :faculty_sponsor,
-      :alignments_in_bam => :aligned
+      id: :id_study_lims,
+      uuid: :uuid_study_lims,
+      sac_sponsor: :faculty_sponsor,
+      alignments_in_bam: :aligned
     )
   end
 end
