@@ -3,10 +3,10 @@ module ResourceTools::Timestamps
 
   included do
     # Ensure that the time stamps are correct whenever a record is updated
-    before_save   { |record| record.recorded_at  = record.checked_time_now }
+    before_save { |record| record.recorded_at = record.checked_time_now }
 
-    delegate :correct_current_time, :to => 'self.class'
-    delegate :checked_time_now, :to => 'self.class'
+    delegate :correct_current_time, to: 'self.class'
+    delegate :checked_time_now, to: 'self.class'
   end
 
   module ClassMethods
