@@ -1,7 +1,6 @@
 require 'spec_helper'
 
 describe Flowcell do
-
   let!(:mock_sample) { create(:sample) }
   let!(:mock_study)  { create(:study)  }
 
@@ -10,9 +9,8 @@ describe Flowcell do
   end
 
   shared_examples_for 'a flowcell' do
-
     it_behaves_like 'maps JSON fields', {
-     :flowcell_id => :id_flowcell_lims
+      flowcell_id: :id_flowcell_lims
     }
 
     it_behaves_like 'ignores JSON fields', [
@@ -21,7 +19,7 @@ describe Flowcell do
     it_behaves_like 'belongs to', [
       :study,
       :sample
-    ], { :lanes => :samples }
+    ], { lanes: :samples }
 
     it_behaves_like 'a nested resource'
   end
@@ -32,76 +30,17 @@ describe Flowcell do
 
     it_behaves_like 'belongs to', [
       :sample
-    ], {:lanes => :controls }
+    ], { lanes: :controls }
 
     it_behaves_like 'a flowcell'
 
-    let(:json) do
-      {
-
-        "flowcell_barcode" => "12345678903",
-        "flowcell_id" => 1123,
-        "forward_read_length" => 222,
-        "reverse_read_length" => 222,
-
-        "updated_at" => "2012-03-11 10:22:42",
-
-        "lanes" => [
-          {
-            "manual_qc" => false,
-            "entity_type" => "library",
-            "position" => 1,
-            "priority" => 1,
-            "id_pool_lims" => "DN324095D A1:H2",
-            "external_release" => true,
-            "purpose" => "standard",
-
-            "samples" => [
-              {
-                "tag_index" => 3,
-                "tag_sequence" => "ATAG",
-                "tag_set_id_lims" => "2",
-                "tag_set_name" => "Sanger_168tags - 10 mer tags",
-                "tag_identifier" => 1,
-                "tag2_sequence" => "GGGG",
-                "tag2_set_id_lims" => "1",
-                "tag2_set_name" => "Tag 2 Set 1",
-                "tag2_identifier" => 1,
-                "pipeline_id_lims" => "Agilent Pulldown",
-                "entity_type" => "library_indexed",
-                "bait_name" => "DDD_V5_plus",
-                "requested_insert_size_from" => 100,
-                "requested_insert_size_to" => 200,
-                "sample_uuid" => "000000-0000-0000-0000-0000000000",
-                "study_uuid" => "000000-0000-0000-0000-0000000001",
-                "cost_code" => "12345",
-                "entity_id_lims" => 12345,
-                "is_r_and_d" => false,
-                "suboptimal" => true
-              }
-            ],
-            "controls" => [
-              {
-                "sample_uuid" => "000000-0000-0000-0000-0000000000",
-                "tag_index" => 168,
-                "entity_type" => "library_indexed_spike",
-                "tag_sequence" => "ATAG",
-                "tag_set_id_lims" => "2",
-                "entity_id_lims" => "12345",
-                "tag_set_name" => "Sanger_168tags - 10 mer tags"
-              }
-            ]
-          }
-        ]
-      }
-    end
+    include_examples 'full flowcell json'
 
     it 'flags all entries as spiked' do
-      Flowcell.all.each {|fc| expect(fc.spiked).to be_true }
+      Flowcell.all.each { |fc| expect(fc.spiked).to be_true }
     end
 
     context 'when update with identical tag indexes' do
-
       let(:example_lims) { 'example' }
 
       let(:updated_json) do
@@ -121,7 +60,6 @@ describe Flowcell do
     end
 
     context 'when update with different tag indexes' do
-
       let(:example_lims) { 'example' }
 
       let(:updated_json) do
@@ -141,7 +79,6 @@ describe Flowcell do
       end
     end
   end
-
 
   context 'without controls or other optional fields' do
     # We have a row for the lane, the sample and the control
@@ -191,13 +128,12 @@ describe Flowcell do
     end
 
     it 'flags all entries as not-spiked' do
-      Flowcell.all.each {|fc| expect(fc.spiked).to be_false }
+      Flowcell.all.each { |fc| expect(fc.spiked).to be_false }
     end
   end
 
   context "a message with clashing samples" do
-
-    let(:expected_identifiers) { 'tag_index, id_flowcell_lims, entity_id_lims, entity_type, position, tag_sequence, tag2_sequence'}
+    let(:expected_identifiers) { 'tag_index, id_flowcell_lims, entity_id_lims, entity_type, position, tag_sequence, tag2_sequence' }
     let(:example_lims) { 'example' }
 
     let(:json) do
@@ -256,9 +192,7 @@ describe Flowcell do
     end
 
     it 'gets rejected' do
-      expect{ described_class.create_or_update_from_json(json, example_lims) }.to raise_error(CompositeResourceTools::InvalidMessage,"Contains two elements with the same composite identifier: combination of #{expected_identifiers} should be unique.")
+      expect { described_class.create_or_update_from_json(json, example_lims) }.to raise_error(CompositeResourceTools::InvalidMessage, "Contains two elements with the same composite identifier: combination of #{expected_identifiers} should be unique.")
     end
-
   end
-
 end
