@@ -1,12 +1,11 @@
 require 'spec_helper'
 
 describe Study do
-
   shared_examples_for 'a study resource' do
     it_behaves_like 'a singular resource'
     it_behaves_like 'maps JSON fields', {
-      :id => :id_study_lims,
-      :sac_sponsor => :faculty_sponsor
+      id: :id_study_lims,
+      sac_sponsor: :faculty_sponsor
     }
 
     it_behaves_like 'ignores JSON fields', [
@@ -16,7 +15,7 @@ describe Study do
     ]
 
     it_behaves_like 'associated with roles' do
-      let(:additional_roles) { [ :data_access_contact, :slf_manager, :lab_manager ] }
+      let(:additional_roles) { [:data_access_contact, :slf_manager, :lab_manager] }
     end
 
     it_behaves_like 'store as boolean', {
