@@ -1,7 +1,6 @@
 shared_examples_for 'maintains roles correctly' do
   after(:each) do
-
-    described_class.create_or_update_from_json(timestamped_json.merge(updated_roles).merge(:updated_at => updated_at),'example')
+    described_class.create_or_update_from_json(timestamped_json.merge(updated_roles).merge(updated_at: updated_at), 'example')
     users_fit_exactly(expected_roles)
   end
 
@@ -17,16 +16,15 @@ end
 shared_examples_for 'associated with roles' do
   def user_with_role(role, index = nil)
     suffix = " #{index}" unless index.nil?
-    { :name => "#{role} name#{suffix}", :email => "#{role} email#{suffix}", :login => "#{role} login#{suffix}" }
+    { name: "#{role} name#{suffix}", email: "#{role} email#{suffix}", login: "#{role} login#{suffix}" }
   end
 
   def users_fit_exactly(roles)
-
     expect(described_class::User.count).to eq(roles.values.flatten.size)
 
     roles.each do |role, expected|
-      found = described_class::User.where(:role => role.to_s).map do |user|
-        Hash[[:name, :email, :login].map { |a| [a,user[a]] }]
+      found = described_class::User.where(role: role.to_s).map do |user|
+        Hash[[:name, :email, :login].map { |a| [a, user[a]] }]
       end
       expect(found).to eq(expected)
     end
@@ -34,14 +32,14 @@ shared_examples_for 'associated with roles' do
 
   let(:additional_roles) { [] }
   let(:originally_created_at) { Time.parse('2012-Mar-16 15:06') }
-  let(:timestamped_json) { json.merge(:created_at => originally_created_at, :updated_at => originally_created_at+1.day) }
+  let(:timestamped_json) { json.merge(created_at: originally_created_at, updated_at: originally_created_at + 1.day) }
 
   context 'for an existing record' do
-    let(:roles)         { { :manager => [ user_with_role(:manager, 1), user_with_role(:manager, 2) ] } }
-    let(:updated_roles) { { :manager => [ user_with_role(:manager, 1), user_with_role(:manager, 2) ] } }
+    let(:roles)         { { manager: [user_with_role(:manager, 1), user_with_role(:manager, 2)] } }
+    let(:updated_roles) { { manager: [user_with_role(:manager, 1), user_with_role(:manager, 2)] } }
 
     before(:each) do
-      described_class.create_or_update_from_json(timestamped_json.merge(roles),'example')
+      described_class.create_or_update_from_json(timestamped_json.merge(roles), 'example')
     end
 
     context 'where the update is classed current it does' do
@@ -58,11 +56,11 @@ shared_examples_for 'associated with roles' do
   end
 
   context 'for new record' do
-    let(:all_role_names) { [ :manager, :follower, :administrator ].concat(additional_roles) }
-    let(:roles) { Hash[all_role_names.map { |role| [role,[user_with_role(role)]] }] }
+    let(:all_role_names) { [:manager, :follower, :administrator].concat(additional_roles) }
+    let(:roles) { Hash[all_role_names.map { |role| [role, [user_with_role(role)]] }] }
 
     before(:each) do
-      described_class.create_or_update_from_json(timestamped_json.merge(roles),'example')
+      described_class.create_or_update_from_json(timestamped_json.merge(roles), 'example')
     end
 
     it 'maintains users' do
