@@ -1,5 +1,5 @@
 class IgnoreableResource
-  def self.create_or_update_from_json(*args)
+  def self.create_or_update_from_json(*_args)
     self.new
   end
 
