@@ -1,9 +1,8 @@
 module SingularResourceTools
-
   extend ActiveSupport::Concern
 
   def latest(other)
-    (updated_values?(other) && ( other.last_updated > last_updated )) ? yield(self) : self
+    (updated_values?(other) && (other.last_updated > last_updated)) ? yield(self) : self
   end
 
   module ClassMethods
