@@ -6,7 +6,7 @@ require 'rspec/rails'
 
 # Requires supporting ruby files with custom matchers and macros, etc,
 # in spec/support/ and its subdirectories.
-Dir[Rails.root.join("spec/support/**/*.rb")].each {|f| require f}
+Dir[Rails.root.join("spec/support/**/*.rb")].each { |f| require f }
 
 RSpec.configure do |config|
   config.use_transactional_fixtures = true
@@ -17,5 +17,4 @@ RSpec.configure do |config|
     ActiveRecord::Migration.maintain_test_schema!
     FactoryGirl.lint
   end
-
 end
