@@ -33,14 +33,12 @@ shared_examples_for 'a singular resource' do |undeletable: true|
   end
 
   context '.create_or_update_from_json' do
-
     context 'from different lims' do
-
       let(:second_lims) { 'example_2' }
 
       before(:each) do
         described_class.create_or_update_from_json(timestamped_json.merge("updated_at" => modified_at), example_lims)
-        described_class.create_or_update_from_json(timestamped_json.merge("updated_at" => modified_at, "uuid" =>'other'), second_lims)
+        described_class.create_or_update_from_json(timestamped_json.merge("updated_at" => modified_at, "uuid" => 'other'), second_lims)
       end
 
       it 'creates multiple records' do
@@ -49,17 +47,14 @@ shared_examples_for 'a singular resource' do |undeletable: true|
     end
 
     context 'without existing records' do
-
       let(:recorded_time) { checked_time_now }
 
       context 'when the record is deleted' do
-
         let(:most_recent_time) { modified_at }
-
 
         before(:each) do
           allow(described_class).to receive(:checked_time_now).and_return(checked_time_now)
-          described_class.send(:create_or_update, attributes.merge("updated_at" =>modified_at,"deleted_at" => modified_at,:id_lims=>example_lims))
+          described_class.send(:create_or_update, attributes.merge("updated_at" => modified_at, "deleted_at" => modified_at, :id_lims => example_lims))
         end
 
         if undeletable
@@ -75,11 +70,10 @@ shared_examples_for 'a singular resource' do |undeletable: true|
       end
 
       context 'when the record is new' do
-
         let(:most_recent_time) { originally_created_at }
         before(:each) do
           allow(described_class).to receive(:checked_time_now).and_return(checked_time_now)
-          described_class.create_or_update_from_json(timestamped_json,example_lims)
+          described_class.create_or_update_from_json(timestamped_json, example_lims)
         end
 
         it_behaves_like 'has only one row'
@@ -87,7 +81,6 @@ shared_examples_for 'a singular resource' do |undeletable: true|
     end
 
     context 'with an existing record' do
-
       before(:each) do
         allow(described_class).to receive(:checked_time_now).and_return(checked_time_then)
         described_class.create_or_update_from_json(timestamped_json.merge("updated_at" => modified_at), example_lims)
@@ -95,7 +88,7 @@ shared_examples_for 'a singular resource' do |undeletable: true|
 
       context 'when the new record is not current' do
         before(:each) do
-          described_class.send(:create_or_update, attributes.merge("updated_at" => modified_at - 2.hours, :id_lims=>example_lims))
+          described_class.send(:create_or_update, attributes.merge("updated_at" => modified_at - 2.hours, :id_lims => example_lims))
         end
 
         it 'only has the current row in the view' do
@@ -105,7 +98,6 @@ shared_examples_for 'a singular resource' do |undeletable: true|
       end
 
       context 'when the fields are unchanged' do
-
         let(:most_recent_time) { modified_at }
         let(:recorded_time) { checked_time_then }
 
@@ -130,7 +122,7 @@ shared_examples_for 'a singular resource' do |undeletable: true|
               # and then update the attribute.
               allow(described_class).to receive(:checked_time_now).and_return(checked_time_now)
               attributes[attribute] = 'changed'
-              described_class.send(:create_or_update, attributes.merge("updated_at" => modified_at,:id_lims =>example_lims))
+              described_class.send(:create_or_update, attributes.merge("updated_at" => modified_at, :id_lims => example_lims))
             end
 
             it_behaves_like 'has only one row'
